fix analyze_fdb
